fix: preserve MacRoman HFS volume names
